若干修改

+++ buss-biz/src/main/java/com/rhtop/buss/biz/service/impl/ContractInfoServiceImpl.java
@@ -94,7 +94,7 @@ public class ContractInfoServiceImpl implements ContractInfoService {
 			tx = txMapper.selectByPrimaryKey(con.getTransactionInfoId());
 			con.setTxAmo(Float.parseFloat(tx.getTxAmo()));
 			con.setCtofPri(tx.getCtofPri());
-			con.setTotPri(Float.parseFloat(tx.getTxAmo())*tx.getCtofPri());
+			con.setTotPri(Double.parseDouble(tx.getTxAmo())*tx.getCtofPri());
 			SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
 			Date d = new Date();
 			String endTime = sdf.format(new Date(d.getTime()+Long.parseLong(tx.getCtofAging())*60*60*1000));
